Time: 83 ms (24.73%), Space: 9.6 MB (69.71%) - LeetHub

+++ 0486-predict-the-winner/0486-predict-the-winner.cpp
@@ -5,8 +5,8 @@ public:
             return 0;
         }
         if(i==j) return nums[i];
-        long long takei=nums[i]+min(solve(i+2,j,nums),solve(i+1,j-1,nums));
-        long long takej=nums[j]+min(solve(i+1,j-1,nums),solve(i,j-2,nums));
+        int takei=nums[i]+min(solve(i+2,j,nums),solve(i+1,j-1,nums));
+        int takej=nums[j]+min(solve(i+1,j-1,nums),solve(i,j-2,nums));
         return max(takei,takej);
     }
     bool predictTheWinner(vector<int>& nums) {
